+++ csrc/ex_include_dirs.h
@@ -6,6 +6,8 @@ void dropDir(void);
 char* getPath(const char* origDir, char* buff);
 char* getCurrentPath(void);
 
+char* getDirName(const char* path, char* buff);
+
 void setCurrentFilename(const char* name);
 char* getCurrentFilename(void);