+++ doc/changes.md
@@ -18,6 +18,8 @@ Changes in '#include's due to file name changes are not included.
 - added 'ID_MINUSLOOP_P'
 - added 'ID_FP_SF_STORE'
 - added 'ID_FP_SF_FETCH'
+- added 'ID_ALTER_PATH_STORE'
+- added 'ID_ALTER_PATH_FETCH'
 - added the 'SFLOAT' macros
 
 ## csrc/excompil.c
@@ -32,10 +34,13 @@ Changes in '#include's due to file name changes are not included.
 - added 'NTIME' word
 - added 'FLOAT' word
 - added '(-LOOP)' word
-- made ffIncludeFile work with ex_include_dirs.c
-- made it so that on error is thrown on BYE
 - added 'SF!' word
 - added 'SF@' word
+- added 'ALTER-PATH!' word
+- added 'ALTER-PATH@' word
+- made ffIncludeFile work with ex_include_dirs.c
+- made ffIncludeFile work with alterPath
+- made it so that on error is thrown on BYE
 
 ## csrc/ex_inner.c
 - added case for token 'ID_INCLUDE_CLIB'
@@ -49,6 +54,8 @@ Changes in '#include's due to file name changes are not included.
 - added case for token 'ID_NTIME'
 - added case for token 'ID_FLOAT'
 - added case for token 'ID_MINUSLOOP_P'
+- added case for token 'ID_ALTER_PATH_STORE'
+- added case for token 'ID_ALTER_PATH_FETCH'
 - replaced FileStream with OpenedFile
 
 ## csrc/exinnrfp.h
@@ -71,6 +78,7 @@ Changes in '#include's due to file name changes are not included.
 
 ## csrc/ex_io.c
 - modified sdFileOpen and sdFileClose to work with OpenedFile struct instead of FILE*
+- added alterPath and alterPathPrev
 
 ## csrc/ex_include_dirs.c
 - includes relative to file
@@ -134,6 +142,7 @@ Changes in '#include's due to file name changes are not included.
 ## fth/ex_floats
 - DFALIGN, SFALIGN, DFALIGNED, SFALIGNED
 - DFFIELD, SFFIELD
+- F> (f greater)
 
 ## fth/mkdicdat.fth
 - added bye, so it doesn't exit to REPL mid compiling